fix: broke the tests further into individual files
